move storage account functions to utils
